+++ NEWS.md
@@ -2,7 +2,8 @@
 
 ## Issues
 
-- drop deprecated `.Names` argument of `structure()` to avoid a warning in r-devel (#728).
+- drop deprecated special names (`.Names`, `.Dim`, `.Dimnames`) in `structure()`
+calls, flagged as a NOTE by r-devel checks (#728).
 
 # officer 0.7.5